DB/Misc: Revert a few factions
Revert faction of Alliance Corpse and Forsaken Corpse to 35 as this is what it is in sniff not 1899/1900 and this was making these "corpse" npcs attack players and other npcs.  Closes #13865

Revert faction of Lord Banehollow since with faction 90 using the Shadowy Potion which should make him friendly to turn in quest of npc name into him but with faction 90 remains hostile even after drinking shadowy potion, this works as should with the original faction of 1434.