DB/Misc: Sealed Vial of Poison
Add condition to bag of fishing treasures to prevent Vialed Seal of Poison from dropping if player has already completed the quest it starts since it this item which was responsible for the bag bugging out and remaining in inventory but unclickable after player removed all visible items.

Also slight update for Mission Plague This as with previous update was failing to display gossip menu options and was just behaving as flight master.