DB/Event: Day of the Dead
* spawn all event creatures and gos.

Respawn creature spawns as some of these where found to be incorrect positions
Spawn all the missing gos for this event.

Ghostly cooking fires are meant to also require orange marigolds to be visible but atm is visible while event is active
The spell 30628 http://www.wowhead.com/spell=30628/arcane-invisibility is not sniffed (in fact in sniff there appears to be no invisibility spell used on these npcs, this is the spell we was already using on these npcs but is not from sniff)

Npcs/gos missing from stormwind as dont have valid sniff for those.