Fix Achievement Defense of the Ancients
The spell that trigger the achievement check, needs to be cast before setting up the variables used for the second round, otherwise the achievement will be granted to the wrong team.
Also, SPELL_END_OF_ROUND needs to be casted on the end of the second round too, this was confirmed in retail.

(cherry picked from commit 5cf794b3797c5c8b911e335fc29ae1e88de3ffea)